refactor contract pages
